backend/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
channels==4.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Django==5.1.4
django-cors-headers==4.6.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
h11==0.14.0
idna==3.10
PyJWT==2.10.1
requests==2.32.3
setuptools==75.6.0
sqlparse==0.5.2
stripe==11.3.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
wheel==0.45.0
